Call super in setUpClass / tearDownClass
Now that we can change the base class to use, we may have something defined in the `setUpClass` and `tearDownClass` of the used base class.

This commit call `super` at the start of the dynamically defined `setUpClass`, just before calling `start_scenario`, end the same at the end of the dynamically defined `tearDownClass`, just after calling `stop_scenario`.